Fixes bug creating a new downloader before update
Previously #clear_download! was being called before updating the
attributes in Eff::Package, for both #save_file= and #version=. This
would cause a new downloader to be created with the old attributes, but
it appeared like everything was ok.

This commit fixes that bug and adds regression tests to guard against
this happening again.